backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
websockets
pydantic
python-dotenv
